Move ansi_codes library to a common location
