+++ base/cvd/cuttlefish/ansi_codes/BUILD.bazel
@@ -1,0 +1,12 @@
+load("//cuttlefish/bazel:rules.bzl", "cf_cc_library")
+
+package(default_visibility = ["//:android_cuttlefish"])
+
+cf_cc_library(
+    name = "ansi_codes",
+    srcs = ["ansi_codes.cc"],
+    hdrs = ["ansi_codes.h"],
+    deps = [
+        "@abseil-cpp//absl/strings",
+    ],
+)
