+++ base/cvd/cuttlefish/host/commands/cvd/cli/commands/monitor/kernel.cc
@@ -22,7 +22,7 @@
 
 #include "absl/strings/str_cat.h"
 
-#include "cuttlefish/host/commands/cvd/cli/commands/monitor/ansi_codes.h"
+#include "cuttlefish/ansi_codes/ansi_codes.h"
 #include "cuttlefish/result/result.h"
 
 namespace cuttlefish {
