+++ base/cvd/cuttlefish/host/commands/cvd/cli/commands/monitor/log_tee.cc
@@ -26,7 +26,7 @@
 #include "absl/strings/str_cat.h"
 #include "absl/strings/str_split.h"
 
-#include "cuttlefish/host/commands/cvd/cli/commands/monitor/ansi_codes.h"
+#include "cuttlefish/ansi_codes/ansi_codes.h"
 #include "cuttlefish/host/commands/cvd/cli/commands/monitor/verbosity.h"
 #include "cuttlefish/result/result.h"
