+++ base/cvd/cuttlefish/host/commands/cvd/cli/commands/monitor/truncate.cc
@@ -23,7 +23,7 @@
 
 #include "absl/strings/str_cat.h"
 
-#include "cuttlefish/host/commands/cvd/cli/commands/monitor/ansi_codes.h"
+#include "cuttlefish/ansi_codes/ansi_codes.h"
 
 namespace cuttlefish {
