+++ base/cvd/cuttlefish/host/commands/cvd/cli/commands/monitor/truncate_test.cc
@@ -21,7 +21,7 @@
 #include "absl/strings/str_cat.h"
 #include "gtest/gtest.h"
 
-#include "cuttlefish/host/commands/cvd/cli/commands/monitor/ansi_codes.h"
+#include "cuttlefish/ansi_codes/ansi_codes.h"
 
 namespace cuttlefish {
