+++ base/cvd/cuttlefish/host/commands/cvd/cli/commands/monitor/verbosity.cc
@@ -18,7 +18,7 @@
 
 #include <string_view>
 
-#include "cuttlefish/host/commands/cvd/cli/commands/monitor/ansi_codes.h"
+#include "cuttlefish/ansi_codes/ansi_codes.h"
 
 namespace cuttlefish {
